task_worker/requirements.txt
requests
pymongo
python-dotenv
loguru
pydantic
